security: Move to new security APIs
A recent commit [1] into the `opi-api` repository revamped the security
APIs. This commit in the `opi-poc` repository migrates the client and
server code to using the new APIs. The current status of the work is
such that we can load and unload connections into strongSwan. Subsequent
work will handle initiating and terminating connections, and flesh out
parameter handling.

Related to #220

[1] https://github.com/opiproject/opi-api/pull/94

Signed-off-by: Kyle Mestery <mestery@mestery.com>