Rewrite the cmdDel function
Need to rewrite the current cmdDel funtion as
doesn't handle correctly the non existing namespace case.
That means that it doesn't reset the VF to its original name when
namespace of Pod doesn't exist.

Also changed the defer function in order to cleanup
the cached Netconf independently of the error value.

Signed-off-by: Dimitrios Markou <dimitrios.markou@est.tech>