feat(opilab-demo): extend "ssh_copy_id" command for individual hosts
Accept another parameter

  ./demo.sh ssh_copy_id "$SSHKEY" $what

where "$what" can be either a hostname or "all".

By default, it now only installs the SSH key on "tgen1" host. This
suffices for most purposes, because

  ./demo.sh exec $host $cmd

will access the host by sshing through "tgen1". So we don't need to
install the ssh key on all hosts. Unless we really want to, which we now
can.

Signed-off-by: Thomas Haller <thaller@redhat.com>